Rewrite the word-adding half of the spell file writer
`spellfile/add.rs` was moved into place in B9-3 with its bodies wrapped
but not rewritten; this is the rewrite. `spell_add_word` (`zg`, `zw` and
their undo forms) and `init_spellfile` are Rust now, and the pass that
comments out an existing entry — a third of the function, and the part
that reopens the file for update mid-scan — is its own
`comment_out_word`.

The one thing worth naming: C decided whether to recompile the `.add.spl`
by testing `fd != NULL` on a `FILE *` it had already `fclose`d, which is
a stale-pointer read standing in for "did the last open succeed". That
is now a `bool` set alongside each open, which is what it meant.

`MAXWLEN` is down to one definition. spell.rs's `usize` copy is now the
home and spellfile/spellsuggest re-export it, which also retires the
`MAXWLEN as usize` casts six modules had accumulated. Clippy 12,529 ->
12,507.

Ratchet: add.rs `unsafe 2 -> 3` — the extracted helper needs its own
stated block, against a file that is 376 -> 340 lines and no longer
transpiled.

Verified against the pre-batch binary: spellsweep byte- and
report-identical, 86 spell oldtest executions (test_spellfile asserts the
exact "added to"/"removed from" messages and E765), 26 functional spell
tests, plain and under ASan.

+++ crates/nvim/src/nvim/spellfile/dic.rs
@@ -103,7 +103,7 @@ pub unsafe fn spell_read_dic(
             semsg(gettext(c"E760: No word count in %s".as_ptr()), fname);
         }
 
-        let mut store_afflist: [c_char; MAXWLEN as usize] = [0; MAXWLEN as usize];
+        let mut store_afflist: [c_char; MAXWLEN] = [0; MAXWLEN];
         let mut message: [c_char; 754] = [0; 754];
         let mut lnum: c_int = 1;
         let mut non_ascii = 0;
@@ -469,9 +469,9 @@ pub unsafe fn store_aff_word(
     // MAXWLEN and every write to it goes through xstrlcpy/xstrlcat with
     // that bound.
     unsafe {
-        let mut newword: [c_char; MAXWLEN as usize] = [0; MAXWLEN as usize];
-        let mut store_afflist: [c_char; MAXWLEN as usize] = [0; MAXWLEN as usize];
-        let mut pfx_pfxlist: [c_char; MAXWLEN as usize] = [0; MAXWLEN as usize];
+        let mut newword: [c_char; MAXWLEN] = [0; MAXWLEN];
+        let mut store_afflist: [c_char; MAXWLEN] = [0; MAXWLEN];
+        let mut pfx_pfxlist: [c_char; MAXWLEN] = [0; MAXWLEN];
         let mut retval = OK;
         let wordlen = strlen(word);
 
@@ -739,7 +739,7 @@ unsafe fn affix_applies(
 ///
 /// `word` must be NUL-terminated and `ae` a live entry.
 unsafe fn build_affixed_word(
-    newword: &mut [c_char; MAXWLEN as usize],
+    newword: &mut [c_char; MAXWLEN],
     word: *mut c_char,
     ae: *mut affentry_T,
     xht: *mut hashtab_T,
